Rename `ProjectUtils` -> `ProjectArguments` and update description
